--- Sources/Swissors/Extensions/Foundation/DateFormatter+Swissors.swift
@@ -1,29 +1,0 @@
-//
-//  DateFormatter.swift
-//  Swissors
-//
-//  Created by Maxim Butin on 27.06.2022.
-//
-
-import Foundation
-
-public extension DateFormatter {
-    
-    static let serverDateFormat: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        return formatter
-    }()
-    
-    static let ISO8601Format: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return formatter
-    }()
-    
-    static let timeZoneFormat: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
-        return formatter
-    }()
-}
