--- Sources/Swissors/Formatters/TimeZoneFormatter.swift
@@ -1,16 +1,0 @@
-//
-//  TimeZoneFormatter.swift
-//  Swissors
-//
-//  Created by Maxim Butin on 27.06.2022.
-//
-
-import Foundation
-
-public class TimZoneFormatter: DefaultDateFormatter {
-    
-    override class var formatter: DateFormatter {
-        let formatter = DateFormatter.timeZoneFormat
-        return formatter
-    }
-}
